+++ src/web/assets/JsApiAsset.php
@@ -12,6 +12,7 @@
 namespace doublesecretagency\mapbox\web\assets;
 
 use craft\web\AssetBundle;
+use doublesecretagency\mapbox\MapboxPlugin;
 
 /**
  * Class JsApiAsset
@@ -32,9 +33,16 @@ class JsApiAsset extends AssetBundle
             MapboxAsset::class,
         ];
 
+        // Whether to use minified JavaScript files
+        $minifyJsFiles = (MapboxPlugin::$plugin->getSettings()->minifyJsFiles ?? false);
+
+        // Optionally use minified files
+        $min = ($minifyJsFiles ? 'min.' : '');
+
+        // Load JS files
         $this->js = [
-            'js/mapbox.js',
-            'js/dynamicmap.js',
+            "js/mapbox.{$min}js",
+            "js/dynamicmap.{$min}js",
         ];
     }